A captured effect that has not finished is not a captured effect
The guard was added on a hunch -- sixteen frames looked close to the longest
thing measured -- and it caught a live one on the first run: proto_c's screen
chirp writes its last register in the fifteenth slot of sixteen, so its NOTE
OFF fell outside the capture window entirely. It replayed as a chirp that
swept up and then held that last note for ever, under whatever came next.

The window is twenty-four frames now and the extractor refuses to ship an
effect still writing on the last one, which is a better rule than a bigger
number: the next dump that needs more will say so rather than losing its tail
quietly. The check listens for thirty frames and asserts every effect, in
every build, has gone quiet by the end of them.

+++ tengen_gba_port/gba/nes_audio.c
@@ -289,12 +289,12 @@ void nes_audio_frame(void) {
  * competing for the same three channels.
  * ----------------------------------------------------------------------- */
 static const uint8_t *g_fx_regs;    /* four bytes a frame, or NULL */
-static uint16_t g_fx_write;         /* one bit a frame: does it write? */
+static uint32_t g_fx_write;         /* one bit a frame: does it write? */
 static uint8_t g_fx_frames;         /* how many frames long */
 static uint8_t g_fx_at;             /* which frame it is on */
 static uint8_t g_fx_channel;
 
-void nes_audio_effect(const uint8_t *regs, uint16_t write, uint8_t frames,
+void nes_audio_effect(const uint8_t *regs, uint32_t write, uint8_t frames,
                        uint8_t channel) {
     if (channel > 1 || frames == 0) { g_fx_regs = 0; return; }
     g_fx_regs = regs;
